Make a response's content size an Integer so that it may be null.  Extract a superclass from FileResponder which will retain the name, and will serve the purpose of responding with a file's content.  The existing class (subclass of the *new* FileResponder) will be the implementation that gets the file content from an actual file in the filesystem, earning its name, FileSystemResponder.

+++ test/fitnesse/http/ChunkedResponseTest.java
@@ -127,7 +127,7 @@ public class ChunkedResponseTest implements ResponseSender {
     public void testContentSize() {
         response.add("12345");
         response.closeAll();
-        assertEquals(5, response.getContentSize());
+        assertEquals(5, (int) response.getContentSize());
     }
 
     @Test
